refactor: dedup resolveChainId + share in-flight helper across server/client
- Remove dead isSecuritizeVaultSync from entities/vault/factory.ts.
- Extract the chainId query-param parsing + validation into
  server/utils/resolve-chain-id.ts. Replaces 3 identical copies in the
  reward handlers plus inline variants in intrinsic-apy, vault-categories,
  and vaults handlers.
- Move createInFlightDedup to utils/in-flight.ts (environment-agnostic).
  server/utils/in-flight.ts re-exports it and keeps the server-only
  scheduleBackgroundRefresh helper. The three rewards composables
  (useBrevis, useFuul, useMerkl) now use the shared dedup instead of
  hand-rolled Map + .finally() boilerplate.
- Update stale 4-min references to 5-min (the warm cadence matched TTL
  since an earlier pass; comments were left behind).
