+++ server/api/rewards/fuul.get.ts
@@ -6,11 +6,10 @@
  * the other. User-specific /claimable-rewards stays direct from the
  * browser — not exposed through this proxy.
  */
-import { createError, getQuery, setResponseHeader } from 'h3'
-import type { H3Event } from 'h3'
+import { createError, setResponseHeader } from 'h3'
 import { createRateLimiter } from '~/server/utils/rate-limit'
 import { logWarn } from '~/server/utils/log'
-import { getEnabledChainIds } from '~/utils/chain-env'
+import { resolveChainId } from '~/server/utils/resolve-chain-id'
 import {
   type FuulProtocol,
   readFuul,
@@ -25,18 +24,6 @@ const rateLimiter = createRateLimiter({
 })
 
 const FUUL_PROTOCOLS: FuulProtocol[] = ['euler', 'euler-looping']
-
-const resolveChainId = (event: H3Event): number => {
-  const raw = getQuery(event).chainId
-  const chainId = Number(raw)
-  if (!Number.isInteger(chainId) || chainId <= 0) {
-    throw createError({ statusCode: 400, statusMessage: 'Invalid chainId' })
-  }
-  if (!getEnabledChainIds().includes(chainId)) {
-    throw createError({ statusCode: 400, statusMessage: 'Unsupported chainId' })
-  }
-  return chainId
-}
 
 const resolveProtocol = async (chainId: number, protocol: FuulProtocol): Promise<unknown> => {
   const cached = readFuul(chainId, protocol)
