+++ server/api/vaults.get.ts
@@ -1,8 +1,8 @@
-import { createError, getQuery, setResponseHeader } from 'h3'
+import { createError, setResponseHeader } from 'h3'
 import { createRateLimiter } from '~/server/utils/rate-limit'
 import { logWarn } from '~/server/utils/log'
+import { resolveChainId } from '~/server/utils/resolve-chain-id'
 import { vaultsCache, refreshChainVaults } from '~/server/utils/vaults-cache'
-import { getEnabledChainIds } from '~/utils/chain-env'
 
 const rateLimiter = createRateLimiter({
   max: 600,
@@ -13,10 +13,11 @@ const rateLimiter = createRateLimiter({
 /**
  * Read-only handler for the chain vaults snapshot.
  *
- * - Steady state: the warm-cache plugin rewrites each entry every 4 min.
- *   Hits always land in a <4 min-old cache. No refresh triggered per request.
- * - After extended warm-cache failure: stale data served up to the 10 min
- *   TTL, then cold-path falls back to a synchronous refresh.
+ * - Steady state: the warm-cache plugin rewrites each entry every 5 min
+ *   via direct `refreshChainVaults()` call (force-refresh). Hits always
+ *   land in a <5 min-old cache. No refresh triggered per request.
+ * - After extended warm-cache failure: stale data served up to the
+ *   staleness ceiling, then cold-path falls back to a synchronous refresh.
  * - Cold start (before first warm cycle completes): handler awaits the
  *   refresh. In-flight dedup in vaults-cache collapses concurrent cold
  *   requests onto a single upstream pass.
@@ -24,13 +25,7 @@ const rateLimiter = createRateLimiter({
 export default defineEventHandler(async (event) => {
   rateLimiter.consume(event)
 
-  const chainId = Number(getQuery(event).chainId)
-  if (!Number.isInteger(chainId) || chainId <= 0) {
-    throw createError({ statusCode: 400, statusMessage: 'Invalid chainId' })
-  }
-  if (!getEnabledChainIds().includes(chainId)) {
-    throw createError({ statusCode: 400, statusMessage: 'Unsupported chainId' })
-  }
+  const chainId = resolveChainId(event)
 
   const cacheKey = String(chainId)
   const cached = vaultsCache.get(cacheKey) ?? vaultsCache.getStale(cacheKey)
